# app/frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MoneyGraph API",
    "version": "1.4.0"
  },
  "paths": {
...
    "/api/v1/meta": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Meta",
        "operationId": "meta_api_v1_meta_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meta"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные входные данные",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Требуется вход",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт состояния прогона",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Слишком большая загрузка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Некорректные параметры запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Слишком много запросов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Сервис временно недоступен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/methodology": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Methodology",
        "operationId": "methodology_api_v1_methodology_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Methodology"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные входные данные",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Требуется вход",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт состояния прогона",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Слишком большая загрузка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Некорректные параметры запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Слишком много запросов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Сервис временно недоступен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldIssue"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "FieldIssue": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "FieldIssue"
      },
...
      "Limits": {
        "properties": {
          "max_upload_mb": {
            "type": "integer",
            "title": "Max Upload Mb"
          },
          "max_zip_members": {
            "type": "integer",
            "title": "Max Zip Members"
          },
          "max_nodes": {
            "type": "integer",
            "title": "Max Nodes"
          },
          "max_transactions": {
            "type": "integer",
            "title": "Max Transactions"
          }
        },
        "type": "object",
        "required": [
          "max_upload_mb",
          "max_zip_members",
          "max_nodes",
          "max_transactions"
        ],
        "title": "Limits"
      },
      "Meta": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version"
          },
          "auth_required": {
            "type": "boolean",
            "title": "Auth Required"
          },
          "llm_enabled": {
            "type": "boolean",
            "title": "Llm Enabled"
          },
          "limits": {
            "$ref": "#/components/schemas/Limits"
          }
        },
        "type": "object",
        "required": [
          "version",
          "auth_required",
          "llm_enabled",
          "limits"
        ],
        "title": "Meta"
      },
      "Methodology": {
        "properties": {
          "roles": {
            "additionalProperties": {
              "$ref": "#/components/schemas/RoleInfo"
            },
            "type": "object",
            "title": "Roles"
          },
          "role_order": {
            "items": {
              "type": "string",
              "enum": [
                "coordinator",
                "consolidator",
                "distributor",
                "transit",
                "terminal",
                "peripheral"
              ]
            },
            "type": "array",
            "title": "Role Order"
          },
          "flags": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Flags"
          },
          "priority_components": {
            "items": {
              "$ref": "#/components/schemas/PriorityComponentInfo"
            },
            "type": "array",
            "title": "Priority Components"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "type": "object",
        "required": [
          "roles",
          "role_order",
          "flags",
          "priority_components",
          "limitations"
        ],
        "title": "Methodology"
      },
...
      "PriorityComponentInfo": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "weight": {
            "type": "number",
            "title": "Weight"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "weight"
        ],
        "title": "PriorityComponentInfo"
      },
...
      "RoleInfo": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "color": {
            "type": "string",
            "title": "Color"
          },
          "rule": {
            "type": "string",
            "title": "Rule"
          }
        },
        "type": "object",
        "required": [
          "label",
          "color",
          "rule"
        ],
        "title": "RoleInfo"
      },
...
    }
  }
}